refactor(confirmations): move EVM account grouping util into confirmations
Keep the helper colocated with confirmation account selection. Shield keeps
a local grouping path to satisfy route isolation.
