+++ ui/pages/shield/transaction-shield/account-selector/account-selector.tsx
@@ -14,6 +14,7 @@ import {
 } from '@metamask/design-system-react';
 import { useSelector } from 'react-redux';
 import classnames from 'clsx';
+import { isEvmAccountType } from '@metamask/keyring-api';
 import {
   Modal,
   ModalBody,
@@ -25,7 +26,6 @@ import {
 import { PreferredAvatar } from '../../../../components/app/preferred-avatar';
 import { AccountSelectorWallet } from '../types';
 import { getWalletsWithAccounts } from '../../../../selectors/multichain-accounts/account-tree';
-import { getEvmAccountsGroupedByWallet } from '../../../../helpers/utils/evm-accounts-grouped-by-wallet';
 import { shortenAddress } from '../../../../helpers/utils/util';
 
 const AccountSelector = ({
@@ -49,13 +49,29 @@ const AccountSelector = ({
   // Group recipients by wallet name
   const accountsGroupedByWallet: Record<string, AccountSelectorWallet> =
     useMemo(() => {
-      return getEvmAccountsGroupedByWallet(wallets).reduce(
+      return Object.values(wallets).reduce(
         (acc, wallet) => {
-          acc[wallet.name] = {
-            id: wallet.id,
-            name: wallet.name,
-            accounts: wallet.accounts,
-          };
+          const walletName = wallet.metadata.name;
+          if (!acc[walletName]) {
+            acc[walletName] = {
+              id: wallet.id,
+              name: wallet.metadata.name,
+              accounts: [],
+            };
+          }
+          Object.values(wallet.groups).forEach((group) => {
+            const evmAccount = group.accounts.find((account) =>
+              isEvmAccountType(account.type),
+            );
+            if (evmAccount) {
+              acc[walletName].accounts.push({
+                id: group.id,
+                name: group.metadata.name,
+                address: evmAccount.address,
+                type: evmAccount.type,
+              });
+            }
+          });
           return acc;
         },
         {} as Record<string, AccountSelectorWallet>,
